Fixed and relocated 'Move' component inside of "Moves"
