+++ react-app/src/app/screens/Game/components/Moves/components/Move/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../../scss/colors';
+@import '../../../../../../../scss/colors.scss';
 
 .move {
   background: $white;
